requirements-dev.txt
-r requirements.txt

binaryornot==0.4.4
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cookiecutter==2.6.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
urllib3==2.4.0
types-python-dateutil==2.9.0.20241206
Pygments==2.19.1
django-browser-reload==1.18.0
django-tailwind==4.0.1